requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Werkzeug==3.0.1
PyMySQL==1.1.0
cryptography==41.0.7
stripe==7.8.0
openai==1.3.7
requests==2.31.0
gunicorn==21.2.0
eventlet==0.33.3
python-engineio==4.8.0
python-socketio==5.10.0
email-validator==2.1.0
Pillow==10.1.0
